refactor: improve files directory
