--- src/app/backing/components/Container/InfoMetrics/index.ts
@@ -1,5 +1,0 @@
-export * from './ABIPopover'
-export * from './AnnualBackersIncentivesMetric'
-export * from './EstimatedRewardsMetric'
-export * from './MetricsContainer'
-export * from './RifRbtcPopover'
