+++ src/app/backing/components/Metrics/EstimatedRewardsMetric.tsx
@@ -1,6 +1,6 @@
 import { Metric, MetricTitle } from '@/app/backing/components/Metric'
 import { useGetBuildersRewards } from '@/app/collective-rewards/rewards/builders/hooks/useGetBuildersRewards'
-import { useEstimatedRewards } from '../hooks/useEstimatedRewards'
+import { useEstimatedRewards } from '../../hooks/useEstimatedRewards'
 import { useHandleErrors } from '@/app/collective-rewards/utils'
 import { COINBASE_ADDRESS } from '@/lib/constants'
 import { tokenContracts } from '@/lib/contracts'
@@ -9,7 +9,7 @@ import KotoQuestionMarkIcon from '@/components/Icons/KotoQuestionMarkIcon'
 import { LoadingSpinner } from '@/components/LoadingSpinner'
 import { Header } from '@/components/TypographyNew'
 import { formatCurrency } from '@/lib/utils'
-import { RifRbtcPopover } from '@/app/backing/components/Container/InfoMetrics'
+import { RifRbtcPopover } from '@/app/backing/components/Popovers/RifRbtcPopover'
 
 export const EstimatedRewardsMetric = () => {
   const tokens = {
